Move header component to separate folder
